# 03 - SPE/ANCE/Scotiabank/Nomina Portal/api/openapi-nomina-portal.yaml
openapi: 3.1.0
...
info:
  title: Portal Empresas Nómina API
  version: 1.0.0
  description: |
    API B2B para que empresas clientes de Scotiabank México gestionen el ciclo de
    nómina bancaria: empleados, nóminas, dispersión vía SPEI y CFDI de nómina.

    Versioning: URI (`/api/v1`). Errores: RFC 9457 (Problem Details).
    Dinero: string-encoded decimal (nunca float). Idempotencia: header `Idempotency-Key`
    obligatorio en operaciones con efecto financiero.
  contact:
    name: Swarm SPE-ANCE-001 · dt-solution-architect
  license:
    name: Internal — Accenture MX / Scotiabank México
servers:
  - url: https://api-nomina.scotiabank.com.mx/api/v1
    description: Producción (placeholder — DATO-REQUERIDO)
  - url: https://api-nomina-qa.scotiabank.com.mx/api/v1
    description: QA
  - url: http://localhost:8080/api/v1
    description: Mock local
...
security:
  - BearerAuth: []

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Autentica un usuario y emite un JWT (mock) / inicia OIDC (prod)
      security: []   # endpoint público
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              operador:
                value:
                  email: operador@empresa.mx
                  password: "Demo1234!"
      responses:
        '200':
          description: Autenticación exitosa (puede requerir 2FA)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                requiere2fa:
                  value:
                    status: PENDING_2FA
                    challengeId: "chg_9a1b"
                    accessToken: null
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/2fa/verify:
    post:
      operationId: verify2fa
      tags: [auth]
      summary: Verifica el segundo factor y emite el JWT de sesión
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorRequest'
            examples:
              codigo:
                value:
                  challengeId: "chg_9a1b"
                  code: "483920"
      responses:
        '200':
          description: 2FA válido — sesión emitida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionToken'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ----------------------------------------------------------- EMPLEADOS (M4)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Credenciales inválidas o token ausente/expirado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ProblemDetails:
      type: object
      description: Error conforme RFC 9457 (Problem Details)
      additionalProperties: true
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: "Código de regla de negocio (RN-xx)" }
        errors:
          type: array
          items:
            type: object
            properties:
              campo: { type: string }
              mensaje: { type: string }
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        status:
          type: string
          enum: [AUTHENTICATED, PENDING_2FA]
        challengeId: { type: string, nullable: true }
        accessToken: { type: string, nullable: true }
    TwoFactorRequest:
      type: object
      additionalProperties: false
      required: [challengeId, code]
      properties:
        challengeId: { type: string }
        code: { type: string }
    SessionToken:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: "Segundos (≤ 3600)" }
        rol:
          type: string
          enum: [ADMIN_EMPRESA, OPERADOR_NOMINA, AUDITOR, ADMIN_SCO]

    # ---- Enums de dominio ----
...